+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Utilities for working with the SCIP solver
 - Helper to check if class attributes have been initialized similar to
     `check_is_fitted` in sklearn
+- Standard validation helpers (`validation_utils`)
 
 ## [0.2.8] - 2018-03-12
 ### Updated